+++ solidblocks-hetzner/modules/rds-postgresql/user_data.sh
@@ -155,6 +155,7 @@ services:
       - "DB_BACKUP_S3_RETENTION_FULL_TYPE=${db_backup_s3_retention_full_type}"
       - "DB_BACKUP_S3_RETENTION_FULL=${db_backup_s3_retention_full}"
       - "DB_BACKUP_S3_RETENTION_DIFF=${db_backup_s3_retention_diff}"
+      - "DB_RESTORE_PITR=${db_restore_pitr}"
       %{~ endif ~}
       %{~ if storage_device_backup != "" ~}
       - "DB_BACKUP_LOCAL=1"