+++ solidblocks-hetzner/test/terraform/instance_s3/main.tf
@@ -26,6 +26,9 @@ module "rds-postgresql-1" {
   backup_s3_retention_full      = 6
   backup_s3_retention_diff      = 2
 
+  postgres_extra_config = "checkpoint_timeout = 30"
+  restore_pitr = var.restore_pitr
+
   databases = [
     { id : "database1", user : "user1", password : "password1" }
   ]